requirements.txt
fastapi==0.115.0
pydantic==2.9.2
PyJWT==2.9.0
PyJWT==2.9.0
uvicorn==0.31.0
